chore: add test species

+++ genomeFiles/run_star_index.sh
@@ -64,8 +64,25 @@ NUM_THREADS=$(nproc)
 GENOME_URL=$(jq -r --arg species "$SPECIES" '.[$species].genome_url' ${PARAMS_JSON_FILE})
 GTF_URL=$(jq -r --arg species "$SPECIES" '.[$species].gtf_url' ${PARAMS_JSON_FILE})
 
+LOCAL_GENOME_FILE=$(jq -r --arg species "$SPECIES" '.[$species].local_genome_file' ${PARAMS_JSON_FILE})
+LOCAL_GTF_FILE=$(jq -r --arg species "$SPECIES" '.[$species].local_gtf_file' ${PARAMS_JSON_FILE})
+
+# We always should have either urls or local files
+if [[ "$LOCAL_GENOME_FILE" != "null" && "$LOCAL_GTF_FILE" != "null" ]];then
+  echo "Found local files."
+  LOCAL_FILES="true"
+  REMOTE_FILES=""
+elif [[ "$GENOME_URL" != "null" && "$GTF_URL" != "null" ]];then
+  echo "Found url files"
+  REMOTE_FILES="true"
+  LOCAL_FILES=""
+else
+ echo "Error: Neither GENOME/GTF Urls nor GENOME/GTF Local files are set."
+ exit 1
+fi
+
 # Check if all required parameters were found
-if [[ -z "$ASSEMBLY_VERSION" || -z "$READ_LENGTH" || -z "$GENOME_URL" || -z "$GTF_URL" ]]; then
+if [[ -z "$ASSEMBLY_VERSION" || -z "$READ_LENGTH" ]]; then
     echo "Error: Parameters for species '$SPECIES' not found in ${PARAMS_JSON_FILE}"
     exit 1
 fi
@@ -74,24 +91,34 @@ fi
 mkdir -p ${OUTPUT_FOLDER}
 cd ${OUTPUT_FOLDER}
 
-# Download and decompress genome DNA fasta file
-echo "Downloading and decompressing genome for ${SPECIES} ${ASSEMBLY_VERSION}..."
-wget -O genome.fa.gz ${GENOME_URL}
-gunzip --force genome.fa.gz
 
-# Download genome annotation GTF file
-GTF_FILENAME="${SPECIES}_${ASSEMBLY_VERSION}_annotations.gtf"
-echo "Downloading and decompressing annotations for ${SPECIES} ${ASSEMBLY_VERSION}..."
-wget -O ${GTF_FILENAME}.gz ${GTF_URL}
-gunzip --force ${GTF_FILENAME}.gz
+if [[ "$REMOTE_FILES" == "true" ]];then
+  # Download and decompress genome DNA fasta file
+  GENOME_FILENAME="genome.fa"
+  echo "Downloading and decompressing genome for ${SPECIES} ${ASSEMBLY_VERSION}..."
+  wget -O genome.fa.gz ${GENOME_URL}
+  gunzip --force genome.fa.gz
+
+  # Download genome annotation GTF file
+  GTF_FILENAME="${SPECIES}_${ASSEMBLY_VERSION}_annotations.gtf"
+  echo "Downloading and decompressing annotations for ${SPECIES} ${ASSEMBLY_VERSION}..."
+  wget -O ${GTF_FILENAME}.gz ${GTF_URL}
+  gunzip --force ${GTF_FILENAME}.gz
+elif [[ "$LOCAL_FILES" == "true" ]];then
+  CURRENT_DIR="${PARENT_SCRIPT_PATH}"
+  GENOME_FILENAME="${CURRENT_DIR}/${LOCAL_GENOME_FILE}"
+  GTF_FILENAME="${CURRENT_DIR}/${LOCAL_GTF_FILE}"
+fi
 
 # Generate Genome Index
 echo "Generating genome index with sjdbOverhang=${SJDB_OVERHANG} and genomeSAindexNbases=${GENOME_SA_INDEX_NBASES}..."
-"${STAR_CMD}" --runThreadN ${NUM_THREADS} --runMode genomeGenerate --genomeDir ./ --genomeFastaFiles genome.fa \
+"${STAR_CMD}" --runThreadN ${NUM_THREADS} --runMode genomeGenerate --genomeDir ./ --genomeFastaFiles ${GENOME_FILENAME} \
      --sjdbGTFfile ${GTF_FILENAME} --sjdbOverhang ${SJDB_OVERHANG} --genomeSAindexNbases ${GENOME_SA_INDEX_NBASES}
 
 # Cleanup: Remove intermediate files
-rm genome.fa
+if [[ -f genome.fa ]];then
+  rm genome.fa
+fi
 rm -rf _STARtmp/
 # Rename GTF to standard name
 mv "${GTF_FILENAME}" annotations.gtf
